Feature/dodanie pobierania z bazy ingredient i recipe

+++ Project/diet/src/main/kotlin/pl/edu/pjwstk/s25236/dietapp/recipe/Recipe.kt
@@ -1,9 +1,10 @@
 package pl.edu.pjwstk.s25236.dietapp.recipe
 
 import pl.edu.pjwstk.s25236.dietapp.ingredient.Ingredient
+import java.math.BigDecimal
 
 class Recipe(
     val id: Long,
-    val ingredients: Map<Ingredient, Double>,
+    val ingredients: Map<Ingredient, BigDecimal>,
     val steps: List<Step>,
 )
