+++ Project/diet/src/main/kotlin/pl/edu/pjwstk/s25236/dietapp/recipe/Step.kt
@@ -2,6 +2,6 @@ package pl.edu.pjwstk.s25236.dietapp.recipe
 
 class Step(
     val id: Long,
-    val position: Int,
+    val position: Long,
     val description: String,
 )
